+++ packages/cli/src/commands/services/status.ts
@@ -1,9 +1,6 @@
 import { homedir } from 'node:os'
-import {
-  getGlobalConfig,
-  getNginxConfigPath,
-  pathExists,
-} from '@denvig/sdk/unsafe'
+import { getGlobalConfig, getNginxConfigPath } from '@denvig/sdk/unsafe'
+import { pathExists } from '@denvig/sdk/utils'
 import { z } from 'zod'
 
 import { Command } from '../../lib/command.ts'